+++ src/Bootstrap/Services.php
@@ -2,7 +2,7 @@
 
 namespace Message\Mothership\Ecommerce\Bootstrap;
 
-use Omnipay\Common\GatewayFactory;
+use Message\Mothership\Ecommerce;
 use Message\Mothership\Ecommerce\Gateway;
 use Message\Mothership\Ecommerce\Statistic;
 use Message\Cog\Bootstrap\ServicesInterface;
@@ -54,6 +54,10 @@ class Services implements ServicesInterface
 			return new \Message\Mothership\Ecommerce\Form\CheckoutRegisterForm($sm);
 		});
 
+		$services['checkout.form.confirm'] = function ($c) {
+			return new \Message\Mothership\Ecommerce\Form\CheckoutConfirmForm($c['gateways'], $c['translator']);
+		};
+
 		$services['product.form.upload_confirm'] = $services->factory(function($c) {
 			return new \Message\Mothership\Ecommerce\Form\Product\CsvUploadConfirm(
 				$c['routing.generator'],
@@ -75,6 +79,10 @@ class Services implements ServicesInterface
 
 		$services['product.form.product_page_publish'] = $services->factory(function($c) {
 			return new \Message\Mothership\Ecommerce\Form\Product\ProductPagePublish;
+		});
+
+		$services['product.form.product_page_create'] = $services->factory(function($c) {
+			return new \Message\Mothership\Ecommerce\Form\Product\ProductPageCreateSingle($c['cms.page.loader'], $c['product.option.loader']);
 		});
 
 		$services['product.page.unit_record.edit'] = $services->factory(function($c) {
@@ -118,7 +126,7 @@ class Services implements ServicesInterface
 		};
 
 		$services['product.page_type.listing'] = function($c) {
-			throw new \LogicException('Service `product.page_type.listing` must be defined within the installation');
+			return null;
 		};
 
 		$services['product.page_type.mapping'] = function($c) {
@@ -139,6 +147,28 @@ class Services implements ServicesInterface
 
 		$services['product.page.brand_validator'] = function($c) {
 			return new \Message\Mothership\Ecommerce\ProductPage\UploadData\BrandValidator($c['product.upload.heading_keys']);
+		};
+
+		// Override from commerce to allow the use of page mappings.
+		$services['product.page_mapper.simple'] = function($c) {
+			$mapper = new \Message\Mothership\Commerce\ProductPageMapper\SimpleMapper(
+				$c['db.query'],
+				$c['cms.page.loader'],
+				$c['cms.page.authorisation'],
+				$c['product.loader'],
+				$c['product.unit.loader']
+			);
+
+			$mapper->setValidFieldNames('product');
+			$mapper->setValidGroupNames(null);
+			$mapper->setValidPageTypes(
+				array_merge(
+					['product'],
+					array_values($c['product.page_type.mapping'])
+				)
+			);
+
+			return $mapper;
 		};
 	}
 
@@ -240,7 +270,30 @@ class Services implements ServicesInterface
 		$services['gateway'] = function($c) {
 			$gateway = $c['cfg']->payment->gateway;
 
+			if (is_array($gateway)) {
+				$gateway = array_shift($gateway);
+			}
+
 			return $c['gateway.collection']->get($gateway);
+		};
+
+		$services['gateways'] = function ($c) {
+			$gateways = $c['cfg']->payment->gateway;
+
+			if (!is_array($gateways)) {
+				$gateways = [$gateways];
+			}
+
+			$collection = $c['gateway.collection'];
+
+			array_walk($gateways, function (&$gateway) use ($collection) {
+				$gateway = $collection->get($gateway);
+			});
+
+			$gateways = new Gateway\Collection($gateways);
+			$gateways->setSort(null);
+
+			return $gateways;
 		};
 	}
 }